requirements.txt
asgiref==3.4.1
astroid==2.7.3
certifi==2021.5.30
charset-normalizer==2.0.6
colorama==0.4.4
cors==1.0.1
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.10.0
django-extensions==3.1.3
django-heroku==0.3.1
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
filelock==3.3.0
future==0.18.2
gevent==21.8.0
greenlet==1.1.2
gunicorn==20.1.0
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
Pillow==8.3.1
platformdirs==2.3.0
psycopg2-binary==2.9.1
PyJWT==2.1.0
pylint==2.10.2
PySocks==1.7.1
pytz==2021.1
random-username==1.0.2
requests==2.26.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.1
tldextract==3.1.2
toml==0.10.2
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.7
whitenoise==5.3.0
wrapt==1.12.1
zope.event==4.5.0
zope.interface==5.4.0
